requirements.txt
Django==3.1.7
django-tinymce==3.2.0
djangorestframework==3.12.2
